#860 CyberTracker - Empty data file created when Import From Device does not find any data
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@2593 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/extensions/org.wcs.smart.cybertracker/src/org/wcs/smart/cybertracker/importer/CyberTrackerImporter.java
@@ -111,8 +111,13 @@ public class CyberTrackerImporter {
 			//move processed files to storage
 			File storageFolder = PdaUtil.getStorageFolder(ca);
 			for (final File file : cxtDataFolder.listFiles()) {
-				if (file.isFile())
-					FileUtils.moveFileToDirectory(file, storageFolder, true);
+				if (file.isFile() && file.getName().toLowerCase().endsWith(".ctx")) { //$NON-NLS-1$
+					if (patrols.isEmpty()) {
+						FileUtils.forceDelete(file);
+					} else {
+						FileUtils.moveFileToDirectory(file, storageFolder, true);
+					}
+				}
 			}
 			
 			return patrols;
